Migrated Player Tracker code to a new dedicated class with slight optimizations
